% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fold_weights}
\alias{add_fold_weights}
\title{Add fold weights to an rset object}
\usage{
add_fold_weights(rset, weights)
}
\arguments{
\item{rset}{An rset object.}

\item{weights}{A numeric vector of weights.}
}
\value{
The rset object with weights added as an attribute.
}
\description{
Add fold weights to an rset object
}
