% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.tune_results}
\alias{autoplot.tune_results}
\title{Plot tuning search results}
\usage{
\method{autoplot}{tune_results}(
  object,
  type = c("marginals", "parameters", "performance"),
  metric = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{object}{A tibble of results from \code{\link[=tune_grid]{tune_grid()}} or \code{\link[=tune_bayes]{tune_bayes()}}.}

\item{type}{A single character value. Choices are \code{"marginals"} (for a plot
of each predictor versus performance; see Details below), \code{"parameters"}
(each parameter versus search iteration), or \code{"performance"} (performance
versus iteration). The latter two choices are only used for \code{\link[=tune_bayes]{tune_bayes()}}.}

\item{metric}{A character vector or \code{NULL} for which metric to plot. By
default, all metrics will be shown via facets.}

\item{width}{A number for the width of the confidence interval bars when
\code{type = "performance"}. A value of zero prevents them from being shown.}

\item{...}{Not currently used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot tuning search results
}
\details{
The parameters are currently represented in their natural units.

When the results of \code{tune_grid()} are used with \code{autoplot()}, it tries to
determine whether a \emph{regular grid} was used.
\subsection{Regular grids}{

For regular grids with one or more numeric tuning parameters, the parameter
with the most unique values is used on the x-axis. If there are categorical
parameters, the first is used to color the geometries. All other parameters
are used in column faceting.

The plot has the performance metric(s) on the y-axis. If there are multiple
metrics, these are row-faceted.

If there are more than five tuning parameters, the "marginal effects" plots
are used instead.
}

\subsection{Irregular grids}{

For space-filling or random grids, a \emph{marginal} effect plot is created. A
panel is made for each numeric parameter so that each parameter is on the
x-axis and performance is on the y-xis. If there are multiple metrics, these
are row-faceted.

A single categorical parameter is shown as colors. If there are two or more
non-numeric parameters, an error is given. A similar result occurs is only
non-numeric parameters are in the grid. In these cases, we suggest using
\code{collect_metrics()} and \code{ggplot()} to create a plot that is appropriate for
the data.
}
}
\examples{
\donttest{
# For grid search:
data("example_ames_knn")

# Plot the tuning parameter values versus performance
autoplot(ames_grid_search, metric = "rmse")


# For iterative search:
# Plot the tuning parameter values versus performance
autoplot(ames_iter_search, metric = "rmse", type = "marginals")

# Plot tuning parameters versus iterations
autoplot(ames_iter_search, metric = "rmse", type = "parameters")

# Plot performance over iterations
autoplot(ames_iter_search, metric = "rmse", type = "performance")
}
}
\seealso{
\code{\link[=tune_grid]{tune_grid()}}, \code{\link[=tune_bayes]{tune_bayes()}}
}
