% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_Bayes.R
\name{Bayes_control}
\alias{Bayes_control}
\title{Control the Bayesian search process}
\usage{
Bayes_control(verbose = FALSE, random_value = 3,
  seed = sample.int(10^5, 1))
}
\arguments{
\item{verbose}{A logical for logging results as they are generated.}

\item{random_value}{The number of iterations with no improvment before an
uncertainty sample is created where a sample with high predicted variance is
chosen.}

\item{seed}{An integer for controlling the random number stream.}
}
\description{
Control the Bayesian search process
}
