% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_632.R
\name{bootstrap_632}
\alias{bootstrap_632}
\title{Bootstrap 632 estimate of performance}
\usage{
bootstrap_632(x)
}
\arguments{
\item{x}{An object of class \code{'tune_results'} that used bootstrap sampling
with the option \code{bootstraps(apparent = TRUE)}.}
}
\value{
A tibble with columns that emulate the structure of
\code{collect_metrics(summarize = TRUE)}. The \code{std_err} column is set to missing
values since the standard error of the 632 estimator is unknown.
}
\description{
This method blends the bootstrap estimate with the resubsitution estimate
to potentially reduce bias.
}
\examples{
\dontshow{if (!tune:::is_cran_check() & tune:::should_run_examples(c("modeldata", "kernlab"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(modeldata)
library(rsample)
library(parsnip)
library(yardstick)

set.seed(1)
tr_dat <- sim_regression(50)
rs <- bootstraps(tr_dat, times = 20, apparent = TRUE)

svm_spec <- svm_rbf(cost = tune()) \%>\% set_mode("regression")

set.seed(2)
svm_res <-
  svm_spec \%>\%
  tune_grid(
    outcome ~ .,
    resamples = rs,
    grid = 2,
    metrics = metric_set(rmse)
  )

collect_metrics(svm_res)
bootstrap_632(svm_res)
\dontshow{\}) # examplesIf}
}
