% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_fold_weights}
\alias{calculate_fold_weights}
\title{Calculate fold weights from fold sizes}
\usage{
calculate_fold_weights(rset)
}
\arguments{
\item{rset}{An rset object.}
}
\value{
A numeric vector of weights proportional to fold sizes.
}
\description{
Calculate fold weights from fold sizes
}
