% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fold_weights}
\alias{check_fold_weights}
\title{Check fold weights in rset objects}
\usage{
check_fold_weights(x)
}
\arguments{
\item{x}{An rset object.}
}
\value{
\code{NULL} invisibly, or error if weights are invalid.
}
\description{
Check fold weights in rset objects
}
\keyword{internal}
