% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-selection.R
\name{choose_metric}
\alias{choose_metric}
\alias{first_metric}
\alias{first_eval_time}
\title{Tools for selecting metrics and evaluation times}
\usage{
choose_metric(x, metric, ..., call = rlang::caller_env())

first_metric(mtr_set)

first_eval_time(mtr_set, metric = NULL, eval_time = NULL)
}
\arguments{
\item{metric}{A character value for which metric is being used.}

\item{mtr_set}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}.}

\item{eval_time}{An optional vector of times to compute dynamic and/or
integrated metrics.}
}
\description{
These are developer-facing functions used to compute and validate choices
for performance metrics. For survival analysis models, there are similar
functions for the evaluation time(s) required for dynamic and/or integrated
metrics.

\code{choose_metric()} is used with functions such as \code{\link[=show_best]{show_best()}} or
\code{\link[=select_best]{select_best()}} where a single valid metric is required to rank models. If
no value is given by the user, the first metric value is used (with a
warning).
}
\keyword{internal}
