% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_predictions}
\alias{collect_predictions}
\alias{collect_metrics}
\title{Obtain and format results produced by tuning}
\usage{
collect_predictions(x, nest_by = "nothing", wflow = NULL)

collect_metrics(x, nest_by = "nothing", wflow = NULL)
}
\arguments{
\item{x}{The results of \code{tune_grid()} or \code{tune_bayes()}. For
\code{collect_predictions()}, the control option \code{save_pred = TRUE} should have
been used.}

\item{nest_by}{A character string to indicate which (if any) columns should
be in the resulting tibble and which should be nested. Values are \code{"nothing"}
(i.e. no nesting), \code{"resamples"} (nest the parameter values), and \code{"parameters"}
(nest the resamples).}

\item{wflow}{The workflow associated with the results (only required for
\code{nest_by = "parameters"}).}
}
\value{
A tibble. The column names depend on the results and the mode of the
model.
}
\description{
Obtain and format results produced by tuning
}
