% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_last_fit}
\alias{control_last_fit}
\title{Control aspects of the last fit process}
\usage{
control_last_fit(verbose = FALSE, event_level = "first", allow_par = FALSE)
}
\arguments{
\item{verbose}{A logical for logging results (other than warnings and errors,
which are always shown) as they are generated during training in a single
R process. When using most parallel backends, this argument typically will
not result in any logging. If using a dark IDE theme, some logging messages
might be hard to see; try setting the \code{tidymodels.dark} option with
\code{options(tidymodels.dark = TRUE)} to print lighter colors.}

\item{event_level}{A single string containing either \code{"first"} or \code{"second"}.
This argument is passed on to yardstick metric functions when any type
of class prediction is made, and specifies which level of the outcome
is considered the "event".}

\item{allow_par}{A logical to allow parallel processing (if a parallel
backend is registered).

If \code{"resamples"}, then tuning will be performed in parallel over resamples
alone. Within each resample, the preprocessor (i.e. recipe or formula) is
processed once, and is then reused across all models that need to be fit.

If \code{"everything"}, then tuning will be performed in parallel at two levels.
An outer parallel loop will iterate over resamples. Additionally, an
inner parallel loop will iterate over all unique combinations of
preprocessor and model tuning parameters for that specific resample. This
will result in the preprocessor being re-processed multiple times, but
can be faster if that processing is extremely fast.

If \code{NULL}, chooses \code{"resamples"} if there are more than one resample,
otherwise chooses \code{"everything"} to attempt to maximize core utilization.

Note that switching between \code{parallel_over} strategies is not guaranteed
to use the same random number generation schemes. However, re-tuning a
model using the same \code{parallel_over} strategy is guaranteed to be
reproducible between runs.}
}
\description{
Control aspects of the last fit process
}
\details{
\code{\link[=control_last_fit]{control_last_fit()}} is a wrapper around \code{\link[=control_resamples]{control_resamples()}} and is meant
to be used with \code{\link[=last_fit]{last_fit()}}.
}
