% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debias_estimate.R
\name{debias_estimate}
\alias{debias_estimate}
\title{Estimate and correct optimization bias}
\usage{
debias_estimate(x)
}
\arguments{
\item{x}{An object with class \code{"tune_results"}. These are produced by
functions such as \code{\link[=tune_grid]{tune_grid()}}, \code{\link[=tune_bayes]{tune_bayes()}}, and so on. There should be
at least two resamples (e.g., validation sets cannot be used).}
}
\value{
A tibble that mimics the output of \code{\link[=show_best]{show_best()}}. There is a row for
each metric and additional columns \code{mean_bias}, \code{std_err_bias}, and
\code{adjusted}. The latter is a corrected value of the \code{mean} column.
}
\description{
Using the same model tuning results to find optimal settings \emph{and} to measure
optimal performance might lead to optimistic performance estimates.
\code{\link[=debias_estimate]{debias_estimate()}} uses the individual resampling statistics to estimate
the bias and adjust the overall resampling estimates.
}
\examples{
debias_estimate(ames_grid_search)
debias_estimate(ames_iter_search)
}
\references{
R J Tibshirani, R Tibshirani "A bias correction for the
minimum error rate in cross-validation," \emph{The Annals of Applied Statistics},
3(2), 822-829, (June 2009)
}
