% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stash_last_result}
\alias{.stash_last_result}
\title{Save most recent results to search path}
\usage{
.stash_last_result(x)
}
\arguments{
\item{x}{An object.}
}
\value{
NULL, invisibly.
}
\description{
Save most recent results to search path
}
\details{
The function will assign \code{x} to \code{.Last.tune.result} and put it in
the search path.
}
