% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_weights.R
\name{.use_case_weights_with_yardstick}
\alias{.use_case_weights_with_yardstick}
\title{Determine if yardstick used case weights}
\usage{
.use_case_weights_with_yardstick(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Not currently used.}
}
\value{
A logical.
}
\description{
This S3 method defines the logical for when a case weight vector should be
passed to the yardstick function and used to measure model performance.
The current logical is that frequency weights are the only situation where
this should occur.
}
\examples{
library(parsnip)
library(dplyr)

frequency_weights(1:10) \%>\%
  .use_case_weights_with_yardstick()

importance_weights(seq(1, 10, by = .1))\%>\%
  .use_case_weights_with_yardstick()

seq(1, 10, by = .1) \%>\%
  .use_case_weights_with_yardstick()
}
