% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/collect.R, R/grid_helpers.R,
%   R/grid_performance.R, R/iteration_results.R,
%   R/loop_over_all_stages-helpers.R, R/loop_over_all_stages.R, R/merge.R,
%   R/parallel.R, R/symbol.R, R/tune_bayes.R, R/tune_grid.R, R/utils.R
\name{check_rset}
\alias{check_rset}
\alias{.check_grid}
\alias{.needs_finalization}
\alias{check_parameters}
\alias{.check_param_objects}
\alias{check_workflow}
\alias{check_metrics}
\alias{check_initial}
\alias{val_class_or_null}
\alias{val_class_and_single}
\alias{.config_key_from_metrics}
\alias{estimate_tune_results}
\alias{.has_preprocessor}
\alias{.has_preprocessor_recipe}
\alias{.has_preprocessor_formula}
\alias{.has_preprocessor_variables}
\alias{.has_spec}
\alias{.set_workflow_spec}
\alias{.set_workflow_recipe}
\alias{metrics_info}
\alias{new_iteration_results}
\alias{.make_static}
\alias{.get_data_subsets}
\alias{.get_config_key}
\alias{.determine_pred_types}
\alias{.loop_over_all_stages}
\alias{.loop_over_all_stages2}
\alias{.update_model}
\alias{.update_recipe}
\alias{.update_parallel_over}
\alias{loop_call}
\alias{get_tune_colors}
\alias{encode_set}
\alias{.is_cataclysmic}
\alias{check_time}
\alias{pull_rset_attributes}
\alias{.set_workflow}
\alias{empty_ellipses}
\alias{is_recipe}
\alias{is_preprocessor}
\alias{is_workflow}
\alias{new_bare_tibble}
\title{Get colors for tune text.}
\usage{
check_rset(x)

.check_grid(grid, workflow, pset = NULL, call = caller_env())

.needs_finalization(x, nms = character(0))

check_parameters(wflow, pset = NULL, data, grid_names = character(0))

.check_param_objects(pset)

check_workflow(x, ..., pset = NULL, check_dials = FALSE, call = caller_env())

check_metrics(x, object)

check_initial(
  x,
  pset,
  wflow,
  resamples,
  metrics,
  eval_time,
  ctrl,
  checks = "grid"
)

val_class_or_null(x, cls = "numeric", where = NULL)

val_class_and_single(x, cls = "numeric", where = NULL)

.config_key_from_metrics(x)

estimate_tune_results(x, ..., col_name = ".metrics")

.has_preprocessor(workflow)

.has_preprocessor_recipe(workflow)

.has_preprocessor_formula(workflow)

.has_preprocessor_variables(workflow)

.has_spec(workflow)

.set_workflow_spec(workflow, spec)

.set_workflow_recipe(workflow, recipe)

metrics_info(x)

new_iteration_results(
  x,
  parameters,
  metrics,
  eval_time,
  eval_time_target,
  outcomes = character(0),
  rset_info,
  workflow
)

.make_static(
  workflow,
  param_info,
  grid,
  metrics,
  eval_time,
  split_args,
  control,
  pkgs = "tune",
  strategy = "sequential",
  data = list(fit = NULL, pred = NULL, cal = NULL)
)

.get_data_subsets(wflow, split, split_args = NULL)

.get_config_key(grid, wflow)

.determine_pred_types(wflow, metrics)

.loop_over_all_stages(resamples, grid, static)

.loop_over_all_stages2(index, resamples, grid, static)

.update_model(grid, object, pset, step_id, nms, ..., source = "model_spec")

.update_recipe(grid, object, pset, step_id, nms, ...)

.update_parallel_over(control, resamples, grid)

loop_call(strategy, framework, opts)

get_tune_colors()

encode_set(x, pset, ..., as_matrix = FALSE)

.is_cataclysmic(x)

check_time(origin, limit)

pull_rset_attributes(x)

.set_workflow(workflow, control)

empty_ellipses(...)

is_recipe(x)

is_preprocessor(x)

is_workflow(x)

new_bare_tibble(x, ..., class = character())
}
\arguments{
\item{x}{An object.}

\item{workflow}{The workflow used to fit the iteration results.}

\item{pset}{A \code{parameters} object.}

\item{wflow}{A \code{workflow} object.}

\item{data}{The training data.}

\item{grid_names}{A character vector of column names from the grid.}

\item{...}{Other options}

\item{check_dials}{A logical for check for a NULL parameter object.}

\item{object}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

A \code{workflow} object.}

\item{resamples}{An \code{rset} object.}

\item{metrics}{A metric set.}

\item{eval_time}{A numeric vector of time points where dynamic event time
metrics should be computed (e.g. the time-dependent ROC curve, etc).}

\item{ctrl}{A \code{control_grid} object.}

\item{cls}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A character vector of
possible classes}

\item{where}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A character string for
the calling function.}

\item{parameters}{A \code{parameters} object.}

\item{outcomes}{A character vector of outcome names.}

\item{rset_info}{Attributes from an \code{rset} object.}

\item{as_matrix}{A logical for the return type.}

\item{origin}{The calculation start time.}

\item{limit}{The allowable time (in minutes).}
}
\description{
These are not intended for use by the general public.
}
\keyword{internal}
