% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{extract_model}
\alias{extract_model}
\title{Convenience functions to extract model (DEPRECATED)}
\usage{
extract_model(x)
}
\arguments{
\item{x}{A fitted workflow object.}
}
\value{
A fitted model.
}
\description{
When extracting the fitted results, the workflow is easily accessible. If
there is only interest in the model, this functions can be used
as a shortcut
}
