% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_recipe}
\alias{extract_recipe}
\alias{extract_model}
\title{Convenience functions to extract model or recipe}
\usage{
extract_recipe(x)

extract_model(x)
}
\arguments{
\item{x}{A fitted workflow object.}
}
\value{
A fitted model or recipe. If a formula is used instead of a recipe,
\code{extract_recipe()} returns \code{NULL}.
}
\description{
When extracting the fitted results, the workflow is easily accessible. If
there is only interest in the recipe or model, these functions can be used
as shortcuts
}
