% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize_model}
\alias{finalize_model}
\alias{finalize_recipe}
\alias{finalize_workflow}
\title{Splice final parameters into objects}
\usage{
finalize_model(x, parameters)

finalize_recipe(x, parameters)

finalize_workflow(x, parameters)
}
\arguments{
\item{x}{A recipe, \code{parsnip} model specification, or workflow.}

\item{parameters}{A list or 1-row tibble of parameter values.}
}
\value{
An updated version of \code{x}.
}
\description{
The \verb{finalize_*} functions take a list or tibble or parameter values and
update objects with those values.
}
