% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_helpers.R
\name{get_metric_time}
\alias{get_metric_time}
\title{Get time for analysis of dynamic survival metrics}
\usage{
get_metric_time(metrics, eval_time)
}
\arguments{
\item{metrics}{A metric set.}

\item{eval_time}{A vector of evaluation times.}
}
\description{
Get time for analysis of dynamic survival metrics
}
\keyword{internal}
