% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_predictions}
\alias{get_predictions}
\title{Obtain and format predictions produced by tuning}
\usage{
get_predictions(x, nest_by = "nothing", wflow = NULL)
}
\arguments{
\item{x}{The results of \code{tune_grid()} or \code{tune_Bayes()} where the predictions
were saved.}

\item{nest_by}{A character string to indicate which (if any) columns should
be in the resulting tibble and which should be nested. Values are \code{"nothing"}
(i.e. no nesting), \code{"resamples"} (nest the parameter values), and \code{"parameters"}
(nest the resamples).}

\item{wflow}{The workflow associated with the results (only required for
\code{nest_by = "parameters"}).}
}
\value{
A tibble. The column names depend on the results and the mode of the
model.
}
\description{
Obtain and format predictions produced by tuning
}
