% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{get_tune_colors}
\alias{get_tune_colors}
\title{Get colors for tune text.}
\usage{
get_tune_colors()
}
\value{
A list of colors that should be used for different types of text.
}
\description{
Get colors for tune text.
}
\keyword{internal}
