% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannah-schedule.R, R/schedule.R
\name{get_tune_schedule}
\alias{get_tune_schedule}
\title{Get tune schedule}
\usage{
get_tune_schedule(grid, wflow)

get_tune_schedule(grid, wflow)
}
\arguments{
\item{grid}{A tibble containing the parameter grid.}

\item{wflow}{A workflow object.}
}
\value{
A schedule object, inheriting from either 'single_schedule',
'grid_schedule', or 'resample_schedule'.

A schedule object, inheriting from either 'single_schedule',
'grid_schedule', or 'resample_schedule'.
}
\description{
Get tune schedule

Get tune schedule
}
