% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid.R
\name{grid_control}
\alias{grid_control}
\title{Control the grid search process}
\usage{
grid_control(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical for logging results as they are generated.}
}
\description{
Control the grid search process
}
