% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_pctl.R
\name{int_pctl_metrics}
\alias{int_pctl_metrics}
\alias{int_pctl_metrics.tune_results}
\title{Bootstrap confidence intervals for performance metrics}
\usage{
int_pctl_metrics(x, ...)

\method{int_pctl_metrics}{tune_results}(
  x,
  metrics = NULL,
  times = 1001,
  alpha = 0.05,
  allow_par = TRUE,
  ...
)
}
\arguments{
\item{x}{A object with class \code{tune_results} where the \code{save_pred = TRUE}
option was used in the control function.}

\item{...}{Not currently used.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}. By default, it uses the same
metrics as the original object.}

\item{times}{The number of bootstrap samples.}

\item{alpha}{Level of significance.}

\item{allow_par}{A logical to allow parallel processing (if a parallel
backend is registered).}
}
\value{
A tibble of metrics with additional columns for \code{.lower} and
\code{.upper}. Note that the \code{.estimate} column is likely to be different from the
results given by \code{\link[=collect_metrics]{collect_metrics()}} since a different estimator is used.
}
\description{
Using out-of-sample predictions, the bootstrap is used to create percentile
confidence intervals.
}
