% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ns.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Quietly load package namespace}
\usage{
load_pkgs(x, ..., infra = TRUE)
}
\arguments{
\item{x}{A character vector of packages.}

\item{infra}{Should base tidymodels packages be loaded as well?}
}
\value{
An invisible NULL.
}
\description{
For one or more packages, load the namespace. This is used during parallel
processing since the different parallel backends handle the package
environments differently.
}
\keyword{internal}
