% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-names.R
\name{outcome_names}
\alias{outcome_names}
\alias{outcome_names.terms}
\alias{outcome_names.formula}
\alias{outcome_names.recipe}
\alias{outcome_names.workflow}
\alias{outcome_names.workflow_variables}
\alias{outcome_names.tune_results}
\title{Determine names of the outcome data in a workflow}
\usage{
outcome_names(x, ...)

\method{outcome_names}{terms}(x, ...)

\method{outcome_names}{formula}(x, ...)

\method{outcome_names}{recipe}(x, ...)

\method{outcome_names}{workflow}(x, ..., call = caller_env())

\method{outcome_names}{workflow_variables}(x, data = NULL, ..., call = caller_env())

\method{outcome_names}{tune_results}(x, ..., call = caller_env())
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to or from other methods (such as \code{data}).}

\item{data}{The training set data (if needed).}
}
\value{
A character string of variable names
}
\description{
Determine names of the outcome data in a workflow
}
\examples{
library(dplyr)
lm(cbind(mpg, wt) ~ ., data = mtcars) \%>\%
  purrr::pluck(terms) \%>\%
  outcome_names()
}
\keyword{internal}
