% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_weights.R
\name{performance_case_weights}
\alias{performance_case_weights}
\alias{performance_case_weights.default}
\alias{performance_case_weights.hardhat_importance_weights}
\alias{performance_case_weights.hardhat_frequency_weights}
\title{Determine if case weights are needed to estimate model performance}
\usage{
performance_case_weights(x, ...)

\method{performance_case_weights}{default}(x, ...)

\method{performance_case_weights}{hardhat_importance_weights}(x, ...)

\method{performance_case_weights}{hardhat_frequency_weights}(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Not currently used.}
}
\value{
A numeric vector or NULL
}
\description{
Determine if case weights are needed to estimate model performance
}
\examples{
freq_wts <- frequency_weights(1:10)
performance_case_weights(freq_wts)

imp_wts <- importance_weights(seq(1, 10, by = .1))
performance_case_weights(imp_wts)

not_wts <- seq(1, 10, by = .1)
performance_case_weights(not_wts)
}
