% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pillar.R
\name{size_sum.tune_results}
\alias{size_sum.tune_results}
\alias{size_sum.resample_results}
\title{Helpers for pillar formatting}
\usage{
size_sum.tune_results(x)

size_sum.resample_results(x)
}
\arguments{
\item{x}{an object}
}
\value{
A character string.
}
\description{
Helpers for pillar formatting
}
\keyword{internal}
