% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_tune_grid.R
\name{pilot_tune_grid}
\alias{pilot_tune_grid}
\alias{pilot_tune_grid.model_spec}
\alias{pilot_tune_grid.workflow}
\title{Run a subset of tuning grid to pre-screen for issues}
\usage{
pilot_tune_grid(object, ...)

\method{pilot_tune_grid}{model_spec}(
  object,
  preprocessor,
  resamples,
  ...,
  param_info = NULL,
  grid = 10,
  metrics = NULL,
  control = control_grid(),
  eval_time = NULL
)

\method{pilot_tune_grid}{workflow}(
  object,
  resamples,
  ...,
  param_info = NULL,
  grid = 10,
  metrics = NULL,
  control = control_grid(),
  eval_time = NULL
)
}
\arguments{
\item{object}{A \code{parsnip} model specification or a \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{...}{Not currently used.}

\item{preprocessor}{A traditional model formula or a recipe created using
\code{\link[recipes:recipe]{recipes::recipe()}}.}

\item{resamples}{An \code{rset()} object.}

\item{param_info}{A \code{\link[dials:parameters]{dials::parameters()}} object or \code{NULL}. If none is given,
a parameters set is derived from other arguments. Passing this argument can
be useful when parameter ranges need to be customized.}

\item{grid}{A data frame of tuning combinations or a positive integer. The
data frame should have columns for each parameter being tuned and rows for
tuning parameter candidates. An integer denotes the number of candidate
parameter sets to be created automatically.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}} or \code{NULL}.}

\item{control}{An object used to modify the tuning process.}

\item{eval_time}{A numeric vector of time points where dynamic event time
metrics should be computed (e.g. the time-dependent ROC curve, etc). The
values must be non-negative and should probably be no greater than the
largest event time in the training set (See Details below).}
}
\value{
An updated version of \code{resamples} with extra list columns for \code{.metrics} and
\code{.notes} (optional columns are \code{.predictions} and \code{.extracts}). \code{.notes}
contains warnings and errors that occur during execution.
}
\description{
\code{\link[=pilot_tune_grid]{pilot_tune_grid()}} finds the extreme values from a pre-defined set of tuning
parameters that correspond to a model or recipe and fits the model on only
one of the resample splits, for purposes of quick error checking.
}
\details{
See \code{\link[=tune_grid]{tune_grid()}} for details of defaults and computation.
}
\seealso{
\code{\link[=tune_grid]{tune_grid()}}, \code{\link[=pilot_fit_resamples]{pilot_fit_resamples()}}
}
