% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_perf_vs_iter}
\alias{plot_perf_vs_iter}
\title{Plot sequential search results}
\usage{
plot_perf_vs_iter(x, metric = NULL, width = max(x$.iter)/75)
}
\arguments{
\item{x}{A tibble of results for \code{tune_Bayes()}.}

\item{metric}{A character vector or \code{NULL} for which outcome to plot.}

\item{width}{A number for the width of the confidence interval bars. A value
of zero prevents them from being shown.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot sequential search results
}
