% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition.R
\name{prob_improve}
\alias{prob_improve}
\alias{exp_improve}
\alias{conf_bound}
\title{Acquisition function for scoring parameter combinations}
\usage{
prob_improve(trade_off = 0, eps = .Machine$double.eps)

exp_improve(trade_off = 0, eps = .Machine$double.eps)

conf_bound(kappa = 1)
}
\arguments{
\item{trade_off}{A number or function that describes the trade-off between
exploitation and exploration.}

\item{eps}{A small constant to avoid division by zero.}

\item{kappa}{A positive number (or function) that corresponds to the multiplier
of the standard deviation in a confidence bounds (e.g. 1.96 in normal-theory
95$\%$ confidence intervals).}
}
\value{
An object of class \code{prob_improve}, \code{exp_improve}, or \code{conf_bounds}
along with an extra class of \code{acquisition_function}.`
}
\description{
These functions can be used to score candidate tuning parameter combinations
as a function of their predicted mean and variation.
}
\details{
The acquisition functions often combine the mean and variance predictions
from the Gaussian process model into an objective to be optimized.

There is a choice between exploitation and exploration.
}
\examples{
prob_improve()
}
