% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dplyr.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dplyr_reconstruct}
\alias{parameters}
\alias{autoplot}
\alias{required_pkgs}
\alias{tune}
\alias{tunable}
\alias{tune_args}
\alias{min_grid}
\alias{augment}
\alias{.get_fingerprint}
\alias{int_pctl}
\alias{extract_spec_parsnip}
\alias{extract_recipe}
\alias{extract_fit_parsnip}
\alias{extract_fit_engine}
\alias{extract_mold}
\alias{extract_preprocessor}
\alias{extract_workflow}
\alias{extract_parameter_set_dials}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dials}{\code{\link[dials:parameters]{parameters()}}}

  \item{dplyr}{\code{\link[dplyr:dplyr_reconstruct]{dplyr_reconstruct()}}}

  \item{generics}{\code{\link[generics:augment]{augment()}}, \code{\link[generics:min_grid]{min_grid()}}, \code{\link[generics:required_pkgs]{required_pkgs()}}, \code{\link[generics:tunable]{tunable()}}, \code{\link[generics:tune_args]{tune_args()}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{hardhat}{\code{\link[hardhat:extract_fit_engine]{extract_fit_engine()}}, \code{\link[hardhat:extract_fit_parsnip]{extract_fit_parsnip()}}, \code{\link[hardhat:extract_mold]{extract_mold()}}, \code{\link[hardhat:extract_parameter_set_dials]{extract_parameter_set_dials()}}, \code{\link[hardhat:extract_preprocessor]{extract_preprocessor()}}, \code{\link[hardhat:extract_recipe]{extract_recipe()}}, \code{\link[hardhat:extract_spec_parsnip]{extract_spec_parsnip()}}, \code{\link[hardhat:extract_workflow]{extract_workflow()}}, \code{\link[hardhat:tune]{tune()}}}

  \item{rsample}{\code{\link[rsample:.get_fingerprint]{.get_fingerprint()}}, \code{\link[rsample:int_pctl]{int_pctl()}}}
}}

