% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample_weights_utils}
\alias{resample_weights_utils}
\alias{.get_resample_weights}
\alias{.create_weight_mapping}
\alias{.weighted_sd}
\alias{.effective_sample_size}
\alias{.validate_resample_weights}
\title{Resampling weights utility functions}
\usage{
.get_resample_weights(x)

.create_weight_mapping(weights, id_names, metrics_data)

.weighted_sd(x, w)

.effective_sample_size(w)

.validate_resample_weights(weights, num_resamples)
}
\arguments{
\item{x}{A tune_results object.}

\item{weights}{Numeric vector of weights.}

\item{id_names}{Character vector of ID column names.}

\item{metrics_data}{The metrics data frame.}

\item{w}{Numeric vector of weights.}

\item{num_resamples}{Integer number of resamples.}
}
\value{
Various return values depending on the function.
}
\description{
These are internal functions for handling variable resampling weights in
hyperparameter tuning.
}
\keyword{internal}
