% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_grid}
\alias{schedule_grid}
\title{Schedule a grid}
\usage{
schedule_grid(grid, wflow)
}
\arguments{
\item{grid}{A tibble containing the parameter grid.}

\item{wflow}{The workflow object for which we schedule the grid.}
}
\value{
A schedule object, inheriting from either 'single_schedule',
'grid_schedule', or 'resample_schedule'.
}
\description{
Schedule a grid
}
\keyword{internal}
