% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_results.R
\name{show_notes}
\alias{show_notes}
\title{Display distinct errors from tune objects}
\usage{
show_notes(x, n = 10)
}
\arguments{
\item{x}{An object of class \code{tune_results}.}

\item{n}{An integer for how many unique notes to show.}
}
\value{
Invisibly, \code{x}. Function is called for side-effects and printing.
}
\description{
Display distinct errors from tune objects
}
