% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarise.tune_results}
\alias{summarise.tune_results}
\title{Estimate performance values}
\usage{
\method{summarise}{tune_results}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}
}
\value{
A tibble with parameter values and summaries of their results. A
column for the average (\code{mean}), the standard error of the mean (\code{std_err}),
and the number of non-missing values (\code{n}). These are computed for each
metric and estimator type.
}
\description{
Estimate performance values
}
