% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw.R
\name{survival_weights}
\alias{survival_weights}
\alias{.survival_weights_graf}
\alias{.survival_weights_roc}
\title{Calculations for two types of IPCW}
\usage{
.survival_weights_graf(
  data,
  truth,
  eval_time,
  model_fit,
  rows = NULL,
  predictors = NULL,
  trunc = 0.05,
  eps = 10^-10
)

.survival_weights_roc(
  data,
  truth,
  eval_time,
  model_fit,
  rows = NULL,
  predictors = NULL,
  trunc = 0.05,
  eps = 10^-10
)
}
\arguments{
\item{data}{A data frame with a column containing a \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{truth}{An unquoted variable name that has the  \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{eval_time}{A vector of non-negative times at which we should
compute the probability of censoring and the corresponding weights. For ROC
analyses, the weights are independent of time but this argument is retained
to keep a consistent API across functions.}

\item{model_fit}{A fitted parsnip model object with a mode of "censored regression".}

\item{rows}{An optional integer vector with length equal to the number of
rows in \code{data} that is used to index the original data. The default is to
use a fresh index on data (i.e. \code{1:nrow(data)}).}

\item{predictors}{Not currently used. A potential future slot for models with
informative censoring based on columns in \code{data}.}

\item{trunc}{A potential lower bound for the probability of censoring to avoid
very large weight values.}
}
\description{
Calculations for two types of IPCW
}
\keyword{internal}
