% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_performance.R
\name{tune-internal-functions}
\alias{tune-internal-functions}
\alias{.estimate_metrics}
\title{Internal functions used by other tidymodels packages}
\usage{
.estimate_metrics(dat, metric, param_names, outcome_name, event_level)
}
\arguments{
\item{dat}{A data set.}

\item{metric}{A metric set.}

\item{param_names}{A character vector of tuning parameter names.}

\item{outcome_name}{A character string for the column of \code{dat} that is the
outcome.}

\item{event_level}{A logical passed from the control function.}

\item{x}{A character vector of package names.}

\item{.expr}{Code to execute.}

\item{...}{Object to pass to the internal \code{tune_log()} function.}

\item{bad_only}{A logical for whether warnings and errors should be caught.}

\item{notes}{Character data to add to the logging.}
}
\description{
These are not to be meant to be invoked directly by users.
}
\keyword{internal}
