% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-package.R
\docType{package}
\name{tune-package}
\alias{tune-package}
\title{tune: Tidy Tuning Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The ability to tune models is important. 'tune' contains functions and classes to be used in conjunction with other 'tidymodels' packages for finding reasonable values of hyper-parameters in models, preprocessing methods, and post-processing steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tune.tidymodels.org/}
  \item \url{https://github.com/tidymodels/tune}
  \item Report bugs at \url{https://github.com/tidymodels/tune/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
