% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_tune_parameters}
\alias{.get_tune_parameters}
\alias{.get_tune_parameter_names}
\alias{.get_extra_col_names}
\alias{.get_tune_metrics}
\alias{.get_tune_metric_names}
\alias{.get_tune_eval_times}
\alias{.get_tune_eval_time_target}
\alias{.get_tune_outcome_names}
\alias{.get_tune_workflow}
\alias{.get_fingerprint.tune_results}
\title{Various accessor functions}
\usage{
.get_tune_parameters(x)

.get_tune_parameter_names(x)

.get_extra_col_names(x)

.get_tune_metrics(x)

.get_tune_metric_names(x)

.get_tune_eval_times(x)

.get_tune_eval_time_target(x)

.get_tune_outcome_names(x)

.get_tune_workflow(x)

\method{.get_fingerprint}{tune_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tune_result}.}
}
\value{
\itemize{
\item \code{.get_tune_parameters()} returns a \code{dials} \code{parameter} object or a tibble.
\item \code{.get_tune_parameter_names()}, \code{.get_tune_metric_names()}, and
\code{.get_tune_outcome_names()} return a character string.
\item \code{.get_tune_metrics()} returns a metric set or NULL.
\item \code{.get_tune_workflow()} returns the workflow used to fit the
resamples (if \code{save_workflow} was set to \code{TRUE} during fitting) or NULL.
}
}
\description{
These functions return different attributes from objects with class
\code{tune_result}.
}
\keyword{internal}
