% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_args.R
\docType{import}
\name{tune_args}
\alias{tune_args}
\alias{tune_args.model_spec}
\alias{tune_args.recipe}
\alias{tune_args.step}
\alias{tune_args.check}
\alias{tune_args.workflow}
\title{Determine arguments tagged for tuning}
\usage{
\method{tune_args}{model_spec}(object, full = FALSE, ...)

\method{tune_args}{recipe}(object, full = FALSE, ...)

\method{tune_args}{step}(object, full = FALSE, ...)

\method{tune_args}{check}(object, full = FALSE, ...)

\method{tune_args}{workflow}(object, ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{tune_args}}}
}}

