% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_nested.R
\name{tune_nested}
\alias{tune_nested}
\alias{tune_nested.model_spec}
\alias{tune_nested.workflow}
\title{Model tuning assessments via nested resampling}
\usage{
tune_nested(object, ...)

\method{tune_nested}{model_spec}(
  object,
  preprocessor,
  resamples,
  fn = "tune_grid",
  param_info,
  metrics,
  .control = control_nested(),
  ...
)

\method{tune_nested}{workflow}(
  object,
  resamples,
  fn = "tune_grid",
  param_info,
  metrics,
  .control = control_nested(),
  ...
)
}
\arguments{
\item{object}{A \code{parsnip} model specification or a \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{...}{Not currently used.}

\item{preprocessor}{A traditional model formula or a recipe created using
\code{\link[recipes:recipe]{recipes::recipe()}}.}

\item{resamples}{An \code{rset()} object.}

\item{param_info}{A \code{\link[dials:parameters]{dials::parameters()}} object or \code{NULL}. If none is given,
a parameters set is derived from other arguments. Passing this argument can
be useful when parameter ranges need to be customized.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}} or \code{NULL}.}

\item{.control}{An object used to modify the tuning process.}

\item{control}{An object used to modify the tuning process.}
}
\value{
An updated version of \code{resamples} with extra list columns for \code{.metrics} and
\code{.notes} (optional columns are \code{.predictions} and \code{.extracts}). \code{.notes}
contains warnings and errors that occur during execution.
}
\description{
\code{\link[=tune_nested]{tune_nested()}} computes
}
\seealso{
\code{\link[=control_grid]{control_grid()}},
\code{\link[=autoplot.tune_results]{autoplot.tune_results()}}, \code{\link[=show_best]{show_best()}}, \code{\link[=select_best]{select_best()}},
\code{\link[=collect_predictions]{collect_predictions()}}, \code{\link[=collect_metrics]{collect_metrics()}}
}
