% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_workflows.R
\name{workflow}
\alias{workflow}
\alias{add_model}
\alias{add_recipe}
\alias{add_formula}
\alias{print.workflow}
\title{Simple workflow functions}
\usage{
workflow()

add_model(object, model)

add_recipe(object, recipe)

add_formula(object, formula)

\method{print}{workflow}(x, ...)
}
\arguments{
\item{object, x}{A workflow object.}

\item{model}{A parsnip model specificatoin.}

\item{recipe}{A recipe object.}

\item{formula}{A basic model formula.}

\item{...}{Not currently used.}
}
\value{
A workflow object.
}
\description{
These functions emulate functions and data structures of the upcoming
workflows package. In the future, these api's may change and these functions
will not be contained in the \code{tune} package.
}
