% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.tune_results}
\alias{autoplot.tune_results}
\title{Plot search results}
\usage{
\method{autoplot}{tune_results}(object, type = c("marginals",
  "parameters", "performance"), metric = NULL, width = NULL, ...)
}
\arguments{
\item{type}{A single character value. Choices are \code{"marginals"} (for a plot
'of each predictor versus performance), \code{"parameters"} (each parameter versus
search iteration), or \code{"performance"} (performance versus iteration). The
latter two choices are only used for \code{tune_Bayes()}.}

\item{metric}{A character vector or \code{NULL} for which outcome to plot.}

\item{width}{A number for the width of the confidence interval bars when
\code{type = "perfomance"}. A value of zero prevents them from being shown.}

\item{x}{A tibble of results form \code{tune_grid()} or \code{tune_Bayes()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot search results
}
\details{
The parameters are currently represented in their natural units.

A single categorical tuning parameter is supported when other numeric
parameters are also in the results. Any number of numeric tuning parameters
can be used.
}
