% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_Bayes.R
\name{Bayes_control}
\alias{Bayes_control}
\title{Control the Bayesian search process}
\usage{
Bayes_control(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical for logging results as they are generated.}
}
\description{
Control the Bayesian search process
}
