% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-selection.R
\name{choose_metric}
\alias{choose_metric}
\alias{first_metric}
\alias{first_eval_time}
\title{Tools for selecting metrics and evaluation times}
\usage{
choose_metric(x, metric)

first_metric(mtr_set)

first_eval_time(mtr_set, metric = NULL, eval_time = NULL)
}
\arguments{
\item{metric}{A character value for which metric is being used.}

\item{mtr_set}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}.}

\item{eval_time}{An optional vector of times to compute dynamic and/or
integrated metrics.}
}
\description{
Tools for selecting metrics and evaluation times
}
\keyword{internal}
