% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-selection.R
\name{choose_metric}
\alias{choose_metric}
\alias{choose_eval_time}
\alias{first_metric}
\alias{first_eval_time}
\alias{.filter_perf_metrics}
\title{Tools for selecting metrics and evaluation times}
\usage{
choose_metric(x, metric, ..., call = rlang::caller_env())

choose_eval_time(x, metric, eval_time = NULL, ..., call = rlang::caller_env())

first_metric(mtr_set)

first_eval_time(mtr_set, metric = NULL, eval_time = NULL)

.filter_perf_metrics(x, metric, eval_time)
}
\arguments{
\item{x}{An object with class \code{tune_results}.}

\item{metric}{A character value for which metric is being used.}

\item{call}{The call to be displayed in warnings or errors.}

\item{eval_time}{An optional vector of times to compute dynamic and/or
integrated metrics.}

\item{mtr_set}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}.}
}
\description{
These are developer-facing functions used to compute and validate choices
for performance metrics. For survival analysis models, there are similar
functions for the evaluation time(s) required for dynamic and/or integrated
metrics.

\code{choose_metric()} is used with functions such as \code{\link[=show_best]{show_best()}} or
\code{\link[=select_best]{select_best()}} where a single valid metric is required to rank models. If
no value is given by the user, the first metric value is used (with a
warning).

For evaluation times, one is only required when the metric type is dynamic
(e.g. \code{\link[yardstick:brier_survival]{yardstick::brier_survival()}} or \code{\link[yardstick:roc_auc_survival]{yardstick::roc_auc_survival()}}). For
these metrics, we require a single numeric value that was originally given
to the function used to produce \code{x} (such as \code{\link[=tune_grid]{tune_grid()}}).

If a time is required and none is given, the first value in the vector
originally given in the \code{eval_time} argument is used (with a warning).
}
\keyword{internal}
