% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_predictions}
\alias{collect_predictions}
\alias{collect_metrics}
\title{Obtain and format results produced by tuning}
\usage{
collect_predictions(x)

collect_metrics(x, summarize = TRUE)
}
\arguments{
\item{x}{The results of \code{\link[=tune_grid]{tune_grid()}} or \code{\link[=tune_bayes]{tune_bayes()}}. For
\code{\link[=collect_predictions]{collect_predictions()}}, the control option \code{save_pred = TRUE} should have
been used.}

\item{summarize}{A logical; should metrics be summarized over resamples
(\code{TRUE}) or return the values for each individual resample.}
}
\value{
A tibble. The column names depend on the results and the mode of the
model.
}
\description{
Obtain and format results produced by tuning
}
\examples{
data("example_ames_knn")

# Summarized over resamples
collect_metrics(ames_grid_search)

# Per-resample values
collect_metrics(ames_grid_search, summarize = FALSE)

# ---------------------------------------------------------------------------

lm_mod <-linear_reg() \%>\% set_engine("lm")
set.seed(93599150)
car_folds <- vfold_cv(mtcars)
ctrl <- control_resamples(save_pred = TRUE)

resampled <- fit_resamples(mpg ~ ., lm_mod, resamples = car_folds, control = ctrl)
collect_predictions(resampled)
}
