% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_grid}
\alias{control_grid}
\alias{control_resamples}
\alias{new_backend_options}
\title{Control aspects of the grid search process}
\usage{
control_grid(
  verbose = FALSE,
  allow_par = TRUE,
  extract = NULL,
  save_pred = FALSE,
  pkgs = NULL,
  save_workflow = FALSE,
  event_level = "first",
  parallel_over = NULL,
  backend_options = NULL,
  workflow_size = 100
)

control_resamples(
  verbose = FALSE,
  allow_par = TRUE,
  extract = NULL,
  save_pred = FALSE,
  pkgs = NULL,
  save_workflow = FALSE,
  event_level = "first",
  parallel_over = NULL,
  backend_options = NULL,
  workflow_size = 100
)

new_backend_options(..., class = character())
}
\arguments{
\item{verbose}{A logical for logging results (other than warnings and errors,
which are always shown) as they are generated during training in a single
R process. When using most parallel backends, this argument typically will
not result in any logging. If using a dark IDE theme, some logging messages
might be hard to see; try setting the \code{tidymodels.dark} option with
\code{options(tidymodels.dark = TRUE)} to print lighter colors.}

\item{allow_par}{A logical to allow parallel processing (if a parallel
backend is registered).}

\item{extract}{An optional function with at least one argument (or \code{NULL})
that can be used to retain arbitrary objects from the model fit object,
recipe, or other elements of the workflow.}

\item{save_pred}{A logical for whether the out-of-sample predictions should
be saved for each model \emph{evaluated}.}

\item{pkgs}{An optional character string of R package names that should be
loaded (by namespace) during parallel processing.}

\item{save_workflow}{A logical for whether the workflow should be appended
to the output as an attribute.}

\item{event_level}{A single string containing either \code{"first"} or \code{"second"}.
This argument is passed on to yardstick metric functions when any type
of class prediction is made, and specifies which level of the outcome
is considered the "event".}

\item{parallel_over}{A single string containing either \code{"resamples"} or
\code{"everything"} describing how to use parallel processing. Alternatively,
\code{NULL} is allowed, which chooses between \code{"resamples"} and \code{"everything"}
automatically.

If \code{"resamples"}, then tuning will be performed in parallel over resamples
alone. Within each resample, the preprocessor (i.e. recipe or formula) is
processed once, and is then reused across all models that need to be fit.

If \code{"everything"}, then tuning will be performed in parallel at two levels.
An outer parallel loop will iterate over resamples. Additionally, an
inner parallel loop will iterate over all unique combinations of
preprocessor and model tuning parameters for that specific resample. This
will result in the preprocessor being re-processed multiple times, but
can be faster if that processing is extremely fast.

If \code{NULL}, chooses \code{"resamples"} if there are more than one resample,
otherwise chooses \code{"everything"} to attempt to maximize core utilization.

Note that switching between \code{parallel_over} strategies is not guaranteed
to use the same random number generation schemes. However, re-tuning a
model using the same \code{parallel_over} strategy is guaranteed to be
reproducible between runs.}

\item{backend_options}{An object of class \code{"tune_backend_options"} as created
by \code{tune::new_backend_options()}, used to pass arguments to specific tuning
backend. Defaults to \code{NULL} for default backend options.}

\item{workflow_size}{A non-negative number (in MB) that is used as a
threshold for a warning regarding the size of the workflow. Only used when
\code{save_workflow = TRUE}.}
}
\value{
An S3 object of class \code{control_grid} (also used for \code{control_bayes},
\code{control_resamples}, \code{control_last_fit}) used to contain the
control settings in a grid search as a list.
}
\description{
Control aspects of the grid search process
}
\details{
For \code{extract}, this function can be used to output the model object, the
recipe (if used), or some components of either or both. When evaluated, the
function's sole argument has a fitted workflow If the formula method is used,
the recipe element will be \code{NULL}.

The results of the \code{extract} function are added to a list column in the
output called \code{.extracts}. Each element of this list is a tibble with tuning
parameter column and a list column (also called \code{.extracts}) that contains
the results of the function. If no extraction function is used, there is no
\code{.extracts} column in the resulting object. See \code{\link[=tune_bayes]{tune_bayes()}} for more
specific details.

Note that for \code{\link[=collect_predictions]{collect_predictions()}}, it is possible that each row of the
original data point might be represented multiple times per tuning
parameter. For example, if the bootstrap or repeated cross-validation are
used, there will be multiple rows since the sample data point has been
evaluated multiple times. This may cause issues when merging the predictions
with the original data.

\code{\link[=control_resamples]{control_resamples()}} is an alias for \code{\link[=control_grid]{control_grid()}} and is meant to be
used with \code{\link[=fit_resamples]{fit_resamples()}}.
}
\section{Hyperparameters and extracted objects}{


When making use of submodels, tune can generate predictions and calculate
metrics for multiple model \code{.config}urations using only one model fit.
However, this means that if a function was supplied to a
\link[=control_grid]{control function's} \code{extract} argument, tune can only
execute that extraction on the one model that was fitted. As a result,
in the \code{collect_extracts()} output, tune opts to associate the
extracted objects with the hyperparameter combination used to
fit that one model workflow, rather than the hyperparameter
combination of a submodel. In the output, this appears like
a hyperparameter entry is recycled across many \code{.config}
entries---this is intentional.

See \url{https://parsnip.tidymodels.org/articles/Submodels.html} to learn
more about submodels.
}

\examples{

library(recipes)
library(rsample)
library(parsnip)
library(workflows)

set.seed(1234)

folds <- vfold_cv(mtcars, v = 5)

recipe <- recipe(mpg ~ ., data = mtcars) |>
  step_normalize(all_numeric_predictors())

# Model with tuning parameters, used for `tune_grid()` and `tune_bayes()`.
spec <- mlp(
  hidden_units = tune(),
  penalty = tune()
) |>
  set_engine("nnet") |>
  set_mode("regression")

wf <- workflow() |>
  add_model(spec) |>
  add_recipe(recipe)

# ---------------------------------------------------------------------------
# control_grid(): Modify tuning process in `tune_grid()`.
ctrl <- control_grid(
  # save out-of-sample predictions
  save_pred = TRUE,
  # print progress while tuning
  verbose = TRUE
)

tune_grid(
  wf,
  resamples = folds,
  grid = 5,
  control = ctrl
)

# ---------------------------------------------------------------------------
# control_bayes(): Modify tuning process in `tune_bayes()`.
ctrl_bayes <- control_bayes(
  save_pred = TRUE,
  verbose = TRUE
)

set.seed(3246)
tune_bayes(
  wf,
  resamples = folds,
  initial = 5,
  iter = 10,
  control = ctrl_bayes
)

# ---------------------------------------------------------------------------
# `fit_resamples()` and `last_fit()` can't use `tune()` placeholders, so for
# `control_resamples()` and `control_last_fit()` we use a model with fixed
# parameters.
# ---------------------------------------------------------------------------
# control_resamples(): Retain out-of-sample predictions and fitted workflows
# in the grid and each fold.
spec_fixed <- linear_reg() |>
  set_engine("lm")

wf_fixed <- workflow() |>
  add_model(spec_fixed) |>
  add_recipe(recipe)

keep_pred <- control_resamples(
  save_pred = TRUE,
  save_workflow = TRUE
)

fit_resamples(
  wf_fixed,
  resamples = folds,
  control = keep_pred
)

# ---------------------------------------------------------------------------
# control_last_fit(): Control last fit process by printing progress.
split <- initial_split(mtcars)

ctrl_last_fit <- control_last_fit(verbose = TRUE)

last_fit(
  wf_fixed,
  split = split,
  control = ctrl_last_fit
)
}
\keyword{internal}
