% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{get_tune_schedule}
\alias{get_tune_schedule}
\title{Get tune schedule}
\usage{
get_tune_schedule(wflow, param, grid)
}
\arguments{
\item{wflow}{A workflow object.}

\item{param}{A dials parameters set.}

\item{grid}{A tibble containing the parameter grid.}
}
\value{
A schedule object, inheriting from either 'single_schedule',
'grid_schedule', or 'resample_schedule'.
}
\description{
Get tune schedule
}
