% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid.R
\name{grid_control}
\alias{grid_control}
\title{Control the grid search process}
\usage{
grid_control(verbose = FALSE, allow_par = TRUE)
}
\arguments{
\item{verbose}{A logical for logging results as they are generated.}

\item{allow_par}{A logical to allow parallel processing (if a parallel
backend is registered).}
}
\description{
Control the grid search process
}
