% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid.R
\name{grid_control}
\alias{grid_control}
\title{Control the grid search process}
\usage{
grid_control(verbose = FALSE, allow_par = TRUE, extract = NULL)
}
\arguments{
\item{verbose}{A logical for logging results as they are generated.}

\item{allow_par}{A logical to allow parallel processing (if a parallel
backend is registered).}

\item{extract}{An optional function with at least one argument (or \code{NULL})
that can be used to retain arbitrary objects from the model fit object,
recipe, or other elements of the workflow.}
}
\description{
Control the grid search process
}
