% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{has_non_par_pkgs}
\alias{has_non_par_pkgs}
\alias{future_installed}
\alias{mirai_installed}
\alias{get_future_workers}
\alias{get_mirai_workers}
\alias{choose_framework}
\alias{get_parallel_seeds}
\alias{eval_mirai}
\alias{par_fns}
\title{Internal functions to help use parallel processing}
\usage{
has_non_par_pkgs(object, control, verbose = FALSE)

future_installed()

mirai_installed()

get_future_workers(verbose)

get_mirai_workers(verbose)

choose_framework(
  object = NULL,
  control = NULL,
  verbose = FALSE,
  default = "mirai"
)

get_parallel_seeds(workers)

eval_mirai(.x, .f, ..., .args)

par_fns(framework)
}
\arguments{
\item{object}{A workflow.}

\item{control}{A control object}

\item{verbose}{A logical for printing}

\item{default}{The default parallel processor.}

\item{workers}{The number of existing workers}

\item{.x}{A list.}

\item{.f}{A function}

\item{..., .args}{Options to pass to other functions.}
}
\description{
Internal functions to help use parallel processing
}
\keyword{internal}
