% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopy.R
\name{loopy}
\alias{loopy}
\title{Iterate over workflow settings}
\usage{
loopy(sched, grid, wflow, tune_id, dat, mtr, eval_time = NULL)
}
\arguments{
\item{sched}{A preprocessing schedule.}

\item{grid}{A parameter grid.}

\item{wflow}{A workflow object.}

\item{tune_id}{A character vector of tuning parameter names.}

\item{dat}{A list containing fit and prediction data.}

\item{mtr}{A metric set.}

\item{eval_time}{Optional. Evaluation time point(s).}
}
\value{
A list containing two elements:
\itemize{
\item \code{metrics}: A tibble of computed metrics grouped by tuning parameters
\item \code{predictions}: A tibble of model predictions
}
}
\description{
Iterate over workflow settings
}
