% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallelism}
\alias{parallelism}
\title{Support for parallel processing in tune}
\description{
Support for parallel backends registered with the \link[foreach:foreach]{foreach}
package was deprecated in tune 1.2.0 in favor of the
\link[future:future]{future} package. tune 1.2.0 was released in March 2024; for
the two years following, the package will raise a warning when:
\enumerate{
\item A parallel backend has been registered with foreach, and
\item No \link[future:plan]{plan} has been specified with future.
}

If parallelism has been configured with both framework, tune will use the
plan specified with future and will not warn. To learn more about plans in
future, see \code{\link[future:plan]{future::plan()}}.
}
