% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_new.R
\name{parallelism}
\alias{parallelism}
\title{Support for parallel processing in tune}
\description{
There are two frameworks that can be used to parallel process your work: the
\link[future:future]{future} package and the \link[mirai:mirai]{mirai} package.
Previously, you could use the \link[foreach:foreach]{foreach} package, but this
has been deprecated as of version 1.2.1 of tune.

By default, no parallelism is used to process models in \pkg{tune}; you have
to opt-in.
\subsection{Using futures}{

You should install the package and choose your flavor of parallelism using
the \link[future:plan]{plan} function. This allows you to specify the number of
worker processes and the specific technology to use.

For example, you can use:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{   library(future)
   plan(multisession, workers = 4)
}\if{html}{\out{</div>}}

and work will be conducted simultaneously (unless there is an exception; see
the section below).

If you had previously used \pkg{foreach}, this would replace your existing
code that probably looked like:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{   library(doBackend)
   registerDoBackend(cores = 4)
}\if{html}{\out{</div>}}

See \code{\link[future:plan]{future::plan()}} for possible options other than \code{multisession}.

Note that \pkg{tune} resets the \emph{maximum} limit of memory of global variables
(e.g., attached packages) to be greater than the default when the package is
loaded. This value can be altered using \code{options(future.globals.maxSize)}.
}

\subsection{Using mirai}{

To set the specific for parallel processing with \pkg{mirai}, the
\code{\link[mirai:daemons]{mirai::daemons()}} functions. The first argument, \code{n}, determines the number
of parallel workers. Using \code{daemons(0)} reverts to sequential processing.

If you want to use \pkg{future} and \pkg{mirai} together, you can install
and load the \pkg{future.mirai} package.
}

\subsection{Exceptions}{

There are a few times when you might specify that you wish to use parallel
processing, but it will revert to sequential execution.
\itemize{
\item Many of the control functions (e.g. \code{\link[=control_grid]{control_grid()}}) have an argument
called \code{allow_par}. If this is set to \code{FALSE}, parallel backends will always
be ignored.
\item Some packages, such as \pkg{rJava} and \pkg{keras} are not compatable with
explict parallelization. If any of these packages are used, sequential
processing occurs.
\item If you specify fewer than two workers, the computations will occur
sequentially.
}
}
}
\references{
https://www.tmwr.org/grid-search#parallel-processing
}
