% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.workflows.R
\name{parameters.workflow}
\alias{parameters.workflow}
\alias{parameters.model_spec}
\alias{parameters.recipe}
\title{Determination of parameter sets for other objects}
\usage{
\method{parameters}{workflow}(x, ...)

\method{parameters}{model_spec}(x, ...)

\method{parameters}{recipe}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Not currently used.}
}
\value{
A parameter set object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These methods have been deprecated in favor of \code{\link[=extract_parameter_set_dials]{extract_parameter_set_dials()}}.
}
\keyword{internal}
