% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{parameters}
\alias{autoplot}
\alias{required_pkgs}
\alias{tunable}
\alias{min_grid}
\alias{augment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dials}{\code{\link[dials]{parameters}}}

  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{min_grid}}, \code{\link[generics]{required_pkgs}}, \code{\link[generics]{tunable}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

