% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ns.R
\name{required_pkgs.model_spec}
\alias{required_pkgs.model_spec}
\alias{required_pkgs.workflow}
\alias{required_pkgs.recipe}
\alias{required_pkgs.step}
\title{Determine packages required by objects}
\usage{
\method{required_pkgs}{model_spec}(x, infra = TRUE, ...)

\method{required_pkgs}{workflow}(x, infra = TRUE, ...)

\method{required_pkgs}{recipe}(x, infra = TRUE, ...)

\method{required_pkgs}{step}(x, ...)
}
\arguments{
\item{x}{An object.}
}
\value{
A character string.
}
\description{
Determine packages required by objects
}
\keyword{internal}
