% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ns.R
\name{required_pkgs}
\alias{required_pkgs}
\title{Determine packages required by objects}
\usage{
required_pkgs(x, ...)
}
\arguments{
\item{x}{An object.}
}
\value{
A character string.
}
\description{
Determine packages required by objects
}
\keyword{internal}
