% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best.R
\name{select_best}
\alias{select_best}
\title{Obtain the best score with parameters produced by tuning}
\usage{
select_best(x, metric = NA, n_top = 1, performance = TRUE,
  maximize = TRUE)
}
\arguments{
\item{x}{The results of \code{tune_grid()} or \code{tune_Bayes()}.}

\item{metric}{A character value. The metric depend on the results and the mode of the
model. e.g. for \code{mode = regression} the metrics could be \code{rmse, rsq, mase} etc.
and for \code{mode = classification} it could be: \code{accuracy, kap, recall, precision, sens, spec, roc_auc, pr_auc, average_precision}
etc. (See
\url{https://tidymodels.github.io/yardstick/articles/metric-types.html} for more details.)
OR a user defined metric for performance during tuning.}

\item{n_top}{number of top results/rows to return.}

\item{performance}{A logical value (TRUE/FALSE) to indicate if columns for the corresponding
performance estimates should also be returned.}

\item{maximize}{A logical value (TRUE/FALSE). This Will be automated later by adding an attribute to the
grid object that has a table that has the appropriate value of maximize for each performance metric
that was requested. Once added, maximize argument will be removed.}
}
\value{
A tibble. The column names depend on the results and the mode of the
model as well as the specified metric type.
}
\description{
Get the best estimator that was chosen by the search, i.e. estimator which gave the highest performance score
(or Parameter setting that gave the best results) on the assessment data.
}
\examples{
\dontrun{
grid_knn <- tune_grid(wflow_obj, rs = cv_splits, control = grid_control(verbose = TRUE))
select_best(grid_knn, metric = "rmse", n_top = 2)
select_best(grid_knn, metric = "rmse")
select_best(grid_knn, metric = "rsq", maximize = FALSE)
}
}
