% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best.R
\name{show_best}
\alias{show_best}
\alias{select_best}
\title{Obtain the numerically best parameters}
\usage{
show_best(x, metric, n_top = 5, maximize = TRUE)

select_best(x, metric, maximize = TRUE)
}
\arguments{
\item{x}{The results of \code{\link[=tune_grid]{tune_grid()}} or \code{\link[=tune_bayes]{tune_bayes()}}.}

\item{metric}{A character value for the metric that will be used to sort
the models. (See
\url{https://tidymodels.github.io/yardstick/articles/metric-types.html} for
more details). Not required if a single metric exists in \code{x}.}

\item{n_top}{An integer for the number of top results/rows to return.}

\item{maximize}{A logical value (TRUE/FALSE).}
}
\value{
A tibble with columns for the parameters. \code{\link[=show_best]{show_best()}} also
includes columns for performance metrics.
}
\description{
\code{\link[=select_best]{select_best()}} finds the tuning parameter combination with the best
performance values. \code{\link[=show_best]{show_best()}} displays the top sub-models and their
performance estimates.
}
\examples{
data("example_ames_knn")

show_best(ames_iter_search, metric = "rmse", maximize = FALSE)

select_best(ames_iter_search, metric = "rsq")
}
