% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{tune-package}
\alias{tune}
\alias{tune-package}
\title{tune: Tidy Tuning Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The ability to tune models is important. 'tune' contains functions and classes to be used in conjunction with other 'tidymodels' packages for finding reasonable values of hyper-parameters in models, pre-processing methods, and post-processing steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tune.tidymodels.org/}
  \item \url{https://github.com/tidymodels/tune}
  \item Report bugs at \url{https://github.com/tidymodels/tune/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
