% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_Bayes.R
\name{tune_Bayes}
\alias{tune_Bayes}
\title{Bayesian optimization of model parameters.}
\usage{
tune_Bayes(object, rs, iter = 10, param_info = NULL, perf = NULL,
  objective = exp_improve(), initial = NULL,
  control = Bayes_control(), ...)
}
\arguments{
\item{object}{A workflow object.}

\item{rs}{A resample object.}

\item{iter}{The maximum number of search iterations.}

\item{param_info}{A \code{dials::param_set()} object or \code{NULL}. If none is given,
a parameters set is derived from the workflow.}

\item{perf}{A \code{yardstick::metric_set()} object containing information on how
models will be evaluated for performance. The first metric in \code{perf} is the
one that will be optimized.}

\item{objective}{A character string for what metric should be optimized or
an acquisition function object.}

\item{initial}{An initial set of results in a tidy format.}

\item{control}{A control object}

\item{...}{Not currently used.}
}
\description{
Bayesian optimization of model parameters.
}
