% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_Bayes.R
\name{tune_Bayes}
\alias{tune_Bayes}
\title{Baeysian optmization of model parameters.}
\usage{
tune_Bayes(object, rs, iter = 10, param_info = NULL, metrics = NULL,
  objective = NULL, initial = NULL, control = Bayes_control(), ...)
}
\arguments{
\item{object}{A workflow object.}

\item{rs}{A resample object.}

\item{iter}{The maximum number of search iterations.}

\item{param_info}{A \code{dials::param_set()} object or \code{NULL}. If none is given,
a parameters set is derived from the workflow.}

\item{metrics}{A \code{metric_set()}}

\item{objective}{A character string for what metric should be optimized or
an aquisition function object.}

\item{initial}{An initial set of results in a tidy format.}

\item{control}{A control object}

\item{...}{Not currently used.}
}
\description{
Baeysian optmization of model parameters.
}
