% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid.R
\name{tune_grid}
\alias{tune_grid}
\title{Model tuning via grid search}
\usage{
tune_grid(object, rs, grid = NULL, perf = NULL,
  control = grid_control())
}
\arguments{
\item{object}{A model workflow object.}

\item{rs}{An \code{rset()} object.}

\item{grid}{A data frame of tuning combinations or \code{NULL}.}

\item{perf}{A \code{yardstick::metric_set()} or \code{NULL}.}

\item{control}{An object used to modify the tuning process.}
}
\value{
A tibble of results.
}
\description{
Model tuning via grid search
}
\details{
If no tuning grid is provided, a semi-random grid (via
\code{dials::grid_latin_hypercube()}) is create with 10 candidate parameter
combinations.

If no metric set is used, one is set. For regression models, the root mean
squared error and coefficient of determination are computed. For classification,
the log-likelihood and overall accuracy are computed.
}
