% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid.R
\name{tune_grid}
\alias{tune_grid}
\alias{tune_grid.default}
\alias{tune_grid.recipe}
\alias{tune_grid.formula}
\alias{tune_grid.workflow}
\title{Model tuning via grid search}
\usage{
tune_grid(object, ...)

\method{tune_grid}{default}(object, ...)

\method{tune_grid}{recipe}(object, model, rs, grid = NULL, perf = NULL,
  control = grid_control(), ...)

\method{tune_grid}{formula}(formula, model, rs, grid = NULL,
  perf = NULL, control = grid_control(), ...)

\method{tune_grid}{workflow}(object, model = NULL, rs, grid = NULL,
  perf = NULL, control = grid_control(), ...)
}
\arguments{
\item{object}{A model workflow or recipe object.}

\item{...}{Not currently used.}

\item{model}{A \code{parsnip} model specification (or \code{NULL} when \code{object} is a
workflow).}

\item{rs}{An \code{rset()} object. the argument \emph{\strong{should be named}}.}

\item{grid}{A data frame of tuning combinations or \code{NULL}.}

\item{perf}{A \code{yardstick::metric_set()} or \code{NULL}.}

\item{control}{An object used to modify the tuning process.}

\item{formula}{A traditional model formula.}
}
\value{
A tibble of results.
}
\description{
Model tuning via grid search
}
\details{
If no tuning grid is provided, a semi-random grid (via
\code{dials::grid_latin_hypercube()}) is create with 10 candidate parameter
combinations.

If no metric set is used, one is set. For regression models, the root mean
squared error and coefficient of determination are computed. For classification,
the log-likelihood and overall accuracy are computed.
}
